feat(P1.3): 오케스트레이터 + finlife 어댑터 + 엔터티 해석
리뷰 수정 4건:
- Rate 저장 형식에 0 패딩. TEXT 사전순이라 "10.0000" < "2.0000"이 참이 되어
  금리 순위가 뒤집히고 있었다. 음수는 예외로 막아 조용히 틀리지 않게 한다.
- ParsedRateRow 금리를 float에서 Decimal로. DECIMAL(7,4) 계약과 일치시킨다.
- make_variant_key에서 amount_min이 0일 때 falsy라 None과 같은 키가 됐다.
- rate_observations.run_id 단독 인덱스 추가. 유니크는 variant_id 선두라
  실행 단위 집계에 쓰이지 않는다. 초기 마이그레이션 재생성.

본작업:
- collectors/finlife/adapter.py: P0에서 검증된 페이지네이션을 httpx로 이식.
  인증키는 환경변수에서만 읽고 request_meta에 넣기 전 마스킹한다.
- services/entity_service.py: source_entity_links 기반 기관·상품·비교단위 해석.
  finlife는 지역을 주지 않으므로 sido/sigungu를 채우지 않는다.
- services/collection_service.py: 명세서 v3 10.1 순서. 실패·차단 시 관측값을
  쓰지 않고 실행 상태만 남긴다.
- cli.py: collect 명령

검증: pytest 89 passed, ruff clean
  실물 fixture 100상품/647옵션 → DB 647관측 관통
  원본 추적 locator 누락 0, raw_artifact_id NULL 0
  재수집 시 정규 엔터티 증가 0, 관측은 실행별 647건
  max_rate NULL 건수가 원본 intr_rate2 NULL 건수와 일치
  금리 왕복 무손실, request_meta에 인증키 0건

+++ src/rate_monitor/domain/schemas.py
@@ -6,6 +6,7 @@
 
 from dataclasses import dataclass, field
 from datetime import date
+from decimal import Decimal
 from typing import Any, Protocol
 
 
@@ -58,8 +59,9 @@ class ParsedRateRow:
     customer_scope: str | None
     availability_scope: str
     rate_scope: str
-    base_rate: float | None
-    max_rate: float | None
+    # DECIMAL(7,4) 계약 (v3 §5.9). float로 낮추면 저장 계층과 타입이 어긋난다.
+    base_rate: Decimal | None
+    max_rate: Decimal | None
     preference_raw: str
     source_row_ref: str
